Implemented the integer spin entry cell editor. Cell editors check the types of the values and report errors if they are incorrect. java.util.Map objects are now presented by the generic perspective. Exceptions and PyExceptions are now presented by the generic perspective.

+++ java/BritefuryJ/DocPresent/Browser/SystemPages/SpinEntryTestPage.java
@@ -62,7 +62,7 @@ public class SpinEntryTestPage extends SystemPage
 		}
 
 
-		public void onSpinEntryValueChanged(IntSpinEntry spinEntry, long value)
+		public void onSpinEntryValueChanged(IntSpinEntry spinEntry, int value)
 		{
 			textElement.setText( String.valueOf( value ) );
 		}
